refactor: player와 dealer의 중복된 코드를 Gamer로 위임 1. 카드의 합을 계산하는 로직 2. card와 이름을 super를 사용해 Gamer에서 초기화
