+++ src/main/java/blackjack/domain/gamer/Dealer.java
@@ -5,33 +5,10 @@ import blackjack.domain.card.Card;
 import java.util.List;
 
 public class Dealer extends Gamer{
-    private static final int TEN = 10;
     private static final int THRESHOLD = 21;
 
-    private List<Card> cards;
-
     public Dealer() {
-        this.cards = initSetting();
-    }
-
-    public int calcScore(Dealer dealer) {
-        int score = dealer.getCards().stream()
-            .map(Card::getDenomination)
-            .mapToInt(Denomination::getValue)
-            .sum();
-
-        long aceCount = dealer.getCards().stream().filter(card -> card.getDenomination().isAce()).count();
-        for(int i = 0; i< aceCount; i++) {
-            score = adjustScore(score);
-        }
-        return score;
-    }
-
-    private int adjustScore(int score) {
-        if(score + TEN <= THRESHOLD) {
-            score += TEN;
-        }
-        return score;
+        super("딜러");
     }
 
     public boolean isBlackJack(Dealer dealer) {
@@ -42,12 +19,4 @@ public class Dealer extends Gamer{
         return calcScore(dealer) > THRESHOLD;
     }
 
-    public List<Card> getCards() {
-        return cards;
-    }
-
-    @Override
-    public String getName() {
-        return "딜러";
-    }
 }
